revactor converts

+++ src/main/java/org/calc/Funcs.java
@@ -5,19 +5,19 @@ public class Funcs {
         return (float)a+(float)b;
     }
     public static float sum(float a, float b){ // +
-        return (float)a+(float)b;
+        return a+b;
     }
     public static float sub(int a, int b){ // -
         return (float)a-(float)b;
     }
     public static float sub(float a, float b){ // -
-        return (float)a-(float)b;
+        return a-b;
     }
     public static float mul(int a, int b){ // *
         return (float)a*(float)b;
     }
     public static float mul(float a, float b){ // *
-        return (float)a*(float)b;
+        return a*b;
     }
     public static float div(int a, int b){ // /
         float c;
